+++ Runtime/Requests/FixedUpdateRateRequest.cs
@@ -1,10 +1,12 @@
+using UniRate.Internals;
+
 namespace UniRate {
 
     public class FixedUpdateRateRequest : RateRequest {
 
         #region <<---------- Initializers ---------->>
 
-        internal FixedUpdateRateRequest(RateManager rateManager, int fixedUpdateRate) : base(rateManager) {
+        internal FixedUpdateRateRequest(RateManagerValueController controller, int fixedUpdateRate) : base(controller) {
             this._fixedUpdateRate = fixedUpdateRate;
         }
 
@@ -21,20 +23,8 @@ namespace UniRate {
         public int FixedUpdateRate => this._fixedUpdateRate;
         private readonly int _fixedUpdateRate;
 
+        protected internal override int Value => this._fixedUpdateRate;
+
         #endregion <<---------- Properties and Fields ---------->>
-
-
-
-
-        #region <<---------- IDisposable ---------->>
-
-        protected override void Dispose(bool disposingManagedResources) {
-            if (this.IsDisposed) return;
-            this.IsDisposed = true;
-            if (this.RateManager == null) return;
-            this.RateManager.CancelFixedUpdateRateRequest(this);
-        }
-
-        #endregion <<---------- IDisposable ---------->>
     }
 }