+++ Runtime/Requests/UpdateRateRequest.cs
@@ -1,10 +1,12 @@
-﻿namespace UniRate {
+﻿using UniRate.Internals;
+
+namespace UniRate {
 
     public class UpdateRateRequest : RateRequest {
 
         #region <<---------- Initializers ---------->>
 
-        internal UpdateRateRequest(RateManager rateManager, int updateRate) : base(rateManager) {
+        internal UpdateRateRequest(RateManagerValueController controller, int updateRate) : base(controller) {
             this._updateRate = updateRate;
         }
 
@@ -21,20 +23,8 @@
         public int UpdateRate => this._updateRate;
         private readonly int _updateRate;
 
+        protected internal override int Value => this._updateRate;
+
         #endregion <<---------- Properties and Fields ---------->>
-
-
-
-
-        #region <<---------- IDisposable ---------->>
-
-        protected override void Dispose(bool disposingManagedResources) {
-            if (this.IsDisposed) return;
-            this.IsDisposed = true;
-            if (this.RateManager == null) return;
-            this.RateManager.CancelUpdateRateRequest(this);
-        }
-
-        #endregion <<---------- IDisposable ---------->>
     }
 }